[UPDATE] updated parameters

+++ dinsargeocoder/dinsargeocoder/terrain_correction.py
@@ -16,7 +16,7 @@ def terrain_correction(flt_product):
     parameters.put('demName', 'SRTM 3Sec')
     parameters.put('demResamplingMethod', 'BILINEAR_INTERPOLATION')
     parameters.put('imgResamplingMethod', 'BILINEAR_INTERPOLATION')
-    parameters.put('pixelSpacingInMeter', float(30))
+    parameters.put('pixelSpacingInMeter', float(50))
     parameters.put('maskOutAreaWithoutElevation', True)
 
     output = snappy.GPF.createProduct("Terrain-Correction", parameters, flt_product)
